requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.3
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-decouple==2.1
django-redis==5.4.0
djangorestframework==3.15.2
kombu==5.3.7
prompt_toolkit==3.0.47
python-dateutil==2.9.0.post0
redis==5.0.7
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
